Longer commit message about Adam F optimizations + minor stuff
- Instead of using wordify() each time two clues are compared, create the full
column of clue_bags (set of non-stopwords in clue) up-front and calculate overlap
of those directly
- use dropna() to remove error-inducing columns where "nan" is the answer line,
because pandas will treat them as null values  (the clues are, I kid you not,
about the Indian flatbread typically spelled 'naan')
- turn the column of bag sizes into a numpy object for incremental speedup
- turn the clue_bag column into a numpy array where each row is a clue and there
are a max of max_clue_length columns. each column is a word within the clue, and
has value -1 if clue is shorter than the length of the longest clue in the df.
this makes it slightly faster than pandas string operations to calculate overlap
between two clues (it's still just set intersection).
- map each unique word across all clue bags to an integer for even faster comparisons
- batch_jaro_winkler builds a "model" object storing every string we want to compare
to; create that model with each unique simple_answer. (note that this model is
returned sorted by length of comparand string, not alphabetically; use numpy to
coerce this output into alphabetical order to match our df) output is a list of
2-tuples whose 0th element is the comparand string and whose 1th element is the
score; collect the 1th element of each tuple
- optimize code applying Pandas masks to reduce redundant calls
- change deletion-marking method; instead of marking rows in the original df with
'_DEL_' (slow), just store a set of row indices to be deleted at the end

Note: there may still be some discrepancies between the reported number of deleted
clues and the actual number; look into that.

Plus, some things I (Matt) did:
-remove unused arguments to remove_redundancies()
-Restore ability to narrow down clues or answers when run from command line
-a bit of style hygiene